backend/requirements.txt
amqp==2.5.0
arrow==0.14.2
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
billiard==3.6.0.0
binaryornot==0.4.4
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
defusedxml==0.6.0
Django==2.2.2
django-environ==0.4.4
django-extensions==2.1.9
django-filters==0.2.1
django-rest-action-permissions==2.0.0
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djoser==1.7.0
flower==0.9.3
future==0.17.1
idna==2.8
importlib-metadata==0.18
Jinja2==2.10.1
jinja2-time==0.2.0
kombu==4.6.3
MarkupSafe==1.1.1
more-itertools==7.1.0
oauthlib==3.0.1
packaging==19.0
Pillow==6.0.0
pkg-resources==0.0.0
pluggy==0.12.0
poyo==0.4.2
py==1.8.0
pycodestyle==2.5.0
PyJWT==1.7.1
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tornado==5.1.1
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
whichcraft==0.5.2
whitenoise==3.3.1
zipp==0.5.2
